Change all the runModal:... methods to return an NSInteger.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@33352 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSSavePanel.m
@@ -1094,7 +1094,7 @@ selectCellWithString: (NSString*)title
   NSCancelButton (if the user clicks the Cancel button).</p>
   <p>See Also: -runModalForDirectory:file:</p>
  */
-- (int) runModal
+- (NSInteger) runModal
 {
   return [self runModalForDirectory: nil file: @""];
 }
@@ -1109,7 +1109,7 @@ selectCellWithString: (NSString*)title
   selected directory or as a last resort, the current working directory.</p>
   <p>See Also: -runModal</p>
  */
-- (int) runModalForDirectory: (NSString*)path file: (NSString*)filename
+- (NSInteger) runModalForDirectory: (NSString*)path file: (NSString*)filename
 {
   [self _setupForDirectory: path file: filename];
   if ([filename length] > 0)
@@ -1117,9 +1117,9 @@ selectCellWithString: (NSString*)title
   return [NSApp runModalForWindow: self];
 }
 
-- (int) runModalForDirectory: (NSString *)path
-			file: (NSString *)filename
-	    relativeToWindow: (NSWindow*)window
+- (NSInteger) runModalForDirectory: (NSString *)path
+                              file: (NSString *)filename
+                  relativeToWindow: (NSWindow*)window
 {
   [self _setupForDirectory: path file: filename];
   if ([filename length] > 0)
